change nsteps and related in DoS to long
git-svn-id: http://zhihaolou.org/svn/neoParSA/trunk@343 a3e5a195-83b0-4c6b-8637-e6d081586ea5

+++ parsa/DoS/DoS.h
@@ -15,7 +15,7 @@ class DoS
 {
 public:
     struct Param {
-        int nSteps;
+        long nSteps;
         double initWeight;
         typename Estimator::Param estParam;
     };
@@ -30,7 +30,7 @@ private:
     unirandom& rnd;
     Estimator estm;
     double weight;
-    int nSteps;
+    long nSteps;
     //double energy;
 };
